# Conflicts:
#	README.md
#	src/main/java/com/malikov/shopsystem/dto/OrderItemDto.java
#	src/main/java/com/malikov/shopsystem/to/OrderTo.java
#	src/main/java/com/malikov/shopsystem/util/OrderUtil.java
#	src/main/java/com/malikov/shopsystem/web/RootController.java
#	src/main/resources/spring/spring-security.xml
#	src/main/webapp/assets/js/orderDatatables.js
#	src/main/webapp/assets/js/productDatatables.js

+++ src/main/java/com/malikov/shopsystem/repository/OrderRepository.java
@@ -8,14 +8,19 @@ import java.util.List;
 
 public interface OrderRepository extends Repository<Order> {
 
-    Collection<Order> getByCustomerId(int customerId);
+    /**
+     * @return orders found by customerId or emply list if not found any
+     */
+    Collection<Order> getByCustomerId(Long customerId);
 
-    Collection<Order> getByProductId(int productId);
+    /**
+     * @return orders found by productId or emply list if not found any
+     */
+    Collection<Order> getByProductId(Long productId);
 
-    void updateStatus(Integer orderId, OrderStatus status);
+    void updateStatus(Long orderId, OrderStatus status);
 
     List<Order> getDatatablePage(int start, int length);
 
     Long getTotalQuantity();
-
 }
